+++ settings/prod.py
@@ -11,6 +11,9 @@ SECRET_KEY = os.environ["SECRET_KEY"]
 
 ALLOWED_HOSTS = ['*']
 
+# Enable Offline Compression for files
+COMPRESS_OFFLINE = True
+
 # Honor the 'X-Forwarded-Proto' header for request.is_secure()
 SECURE_PROXY_SSL_HEADER = ('HTTP_X_FORWARDED_PROTO', 'https')
 
@@ -45,3 +48,7 @@ AWS_SECRET_ACCESS_KEY = os.environ['AWS_SECRET_ACCESS_KEY']
 
 # The S3 bucket used to store uploaded files.
 AWS_S3_BUCKET_NAME = os.environ['S3_BUCKET']
+
+# Define Storage For Thumbnails
+# Use same storage as the one defined for production
+THUMBNAIL_DEFAULT_STORAGE = DEFAULT_FILE_STORAGE