+++ CHANGELOG.md
@@ -2,3 +2,10 @@
 
 * Initial version release
 * Production images stored in S3
+
+## v0.2.0
+
+* Enable Offline Compression for JS/CSS in production
+* Add Sharethis to product detail page
+* Add Google Analytics
+* Generate and display thumbnails on the homepage